rag-mcp/pyproject.toml
[project]
name = "life-rag-mcp"
version = "0.1.0"
description = "RAG MCP server for life repository semantic search"
readme = {text = "", content-type = "text/plain"}
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "mcp>=1.2.0",
    "pyyaml>=6.0",
    "sqlite-vec>=0.1.6",
]

[project.scripts]
life-rag-mcp = "life_rag_mcp.server:main"
life-rag-index = "life_rag_mcp.indexer:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.uv.sources]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/life_rag_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
